this could be final

+++ controllers/products.js
@@ -1,5 +1,5 @@
 const {addProductdb, getProductsdb, editProductdb, grtProductdb, deleteProductdb} = require("../models/products/products")
-const {addTransactiondb,getTransactiondb,getTransactionsdb,updateTransactiondb,deleteTransactiondb} = require("../models/transactions")
+const {addTransactiondb} = require("../models/transactions")
 
 const addProduct = async (req, res) => {
   try {
@@ -8,29 +8,31 @@ const addProduct = async (req, res) => {
       description, 
       category, 
       price, 
-      quantity 
+      quantity,
+      userId
     } = req.body;
 
-    if (!name || !category || !description || !price || !quantity) {
-      return res.status(400).json({ 
-        error: 'Missing fields' 
-      });
+    if (!name.trim() || !category.trim() || !description.trim() || !price || !quantity) {
+      return res.status(400).send('Missing fields');
+    }
+    if (!userId) {
+      return res.status(401).send('You are not authorized');
     }
     if (price <=0 || quantity<=0 || isNaN(price) || isNaN(quantity)) {
-      return res.status(400).json({ 
-        error: 'Invalid data for quantity or price' 
-      });
+      return res.status(400).send('Invalid data for quantity or price' );
     }
-    const products = await getProductsdb('00')
+    const products = await getProductsdb(userId)
 
     const existingProduct = products.find(product=>product.name===name)
     
     if(existingProduct) {
       const transaction = {
+        userId,
         productId:existingProduct.productId,
-        addedQuantity:quantity,
+        addedQuantity:Number(quantity),
         subtractedQuantity:0,
-        soldQuantity:0
+        soldQuantity:0,
+        createdAt: Date.now()
       }
       const newQuantity = parseInt(existingProduct.quantity) + parseInt(quantity)
       const update = await editProductdb(existingProduct.productId,{quantity:newQuantity,price,category,description})
@@ -40,6 +42,7 @@ const addProduct = async (req, res) => {
     
     
     const product = {
+      userId,
       name,
       description,
       category,
@@ -50,37 +53,42 @@ const addProduct = async (req, res) => {
 
     const response = await addProductdb(product)
     const transaction = {
+        userId,
         productId:response.productId,
-        addedQuantity:quantity,
+        addedQuantity:Number(quantity),
         subtractedQuantity:0,
-        soldQuantity:0
+        soldQuantity:0,
+        createdAt: Date.now()
       }
     await addTransactiondb(transaction)
-    res.status(201).json(response.status);
+    return res.status(201).json(response.status);
   } catch (error) {
-    res.status(500).json({ 
-      error: 'Failed to create product', 
-      details: error.message 
-    });
+    return res.status(500).send(err.message)
   }
 }
 const getProducts = async (req, res) =>{
+  const {userId} = req.params
+  if(!userId.trim()) return res.status(401).send('You are not authenticated')
   try{
-    const products = await getProductsdb('oo')
+    const products = await getProductsdb(userId)
     res.json(products)
+    return 
   } catch(err){
-    console.log(err)
+    return res.status(500).send(err.message)
   }
 }
 const deleteProduct = async (req, res) =>{
   try{
-    const {productId} = req.query
+    const {productId,userId} = req.query
+    if(!productId || !userId) return res.status(400).send('Missing fields')
     const productToDelete = await grtProductdb(productId)
     if(!productToDelete) return res.status((404)).send("product not found")
+    if(productToDelete?.userId !== userId) return res.status((401)).send("You are not authorized to delete the product")
     const deleteStatus = await deleteProductdb(productId)
     res.json(deleteStatus)
+    return
   } catch(err){
-    console.log(err)
+    return res.status(500).send(err.message)
   }
 }
 const updateProduct = async (req, res) =>{
@@ -91,21 +99,23 @@ const updateProduct = async (req, res) =>{
       category, 
       price, 
       quantity,
-      productId 
+      productId,
+      userId 
     } = req.body;
-    if (!name || !category || !description || !price || !quantity) {
-      return res.status(400).json({ 
-        error: 'Missing fields' 
-      });
+    if (!name.trim() || !category.trim() || !description.trim() || !price || !quantity || !productId.trim()) {
+      return res.status(400).send('Missing fields' 
+      );
+    }
+    if (!userId ) {
+      return res.status(401).send('You are not authorized');
     }
     if (price <=0 || quantity<=0 || isNaN(price) || isNaN(quantity)) {
-      return res.status(400).json({ 
-        error: 'Invalid data for quantity or price' 
-      });
+      return res.status(400).send('Invalid data for quantity or price');
     }
     
     const productToUpdate = await grtProductdb(productId)
     if(!productToUpdate) return res.status((404)).send("product not found")
+    if(productToUpdate?.userId !== userId) return res.status((401)).send("You are not authorized to update the product")
     const product = {
       name,
       description,
@@ -114,10 +124,63 @@ const updateProduct = async (req, res) =>{
       quantity: parseInt(quantity),
       updatedAt: Date.now().toString()
     };
+    
+    if(Number(quantity) < Number(productToUpdate.quantity)) {
+      const newQuantity = Number(productToUpdate.quantity) - Number(quantity)
+      const transaction = {
+        userId,
+        productId:productToUpdate.productId,
+        addedQuantity:0,
+        subtractedQuantity:newQuantity,
+        soldQuantity:0,
+        createdAt: Date.now()
+      }
+    await addTransactiondb(transaction)
+    }
+    if(Number(quantity) > Number(productToUpdate.quantity)) {
+      const newQuantity = Number(quantity) - Number(productToUpdate.quantity)
+      const transaction = {
+        userId,
+        productId:productToUpdate.productId,
+        addedQuantity:newQuantity,
+        subtractedQuantity:0,
+        soldQuantity:0,
+        createdAt: Date.now()
+      }
+      await addTransactiondb(transaction)
+    }
     const updateStatus = await editProductdb(productId,product)
-    res.json(updateStatus)
+    return res.status(200).json(updateStatus)
   } catch(err){
-    console.log(err)
+    return res.status(500).send(err.message)
   }
 }
-module.exports = {addProduct,getProducts,deleteProduct,updateProduct}
+const sellProduct = async (req, res) =>{
+  try{
+    const {productId,userId, quantity} = req.body
+    if(!productId.trim() || !userId.trim() || !quantity.trim()) return res.status(400).send('Missing fields')
+
+    if(Number(quantity) <=0 || isNaN(quantity)) return res.status(400).send('Invalid Quantity')
+    
+    const productToSell = await grtProductdb(productId)
+
+    if(!productToSell) return res.status((404)).send("product not found")
+    if(productToSell?.userId !== userId) return res.status((401)).send("You are not authorized to delete the product")
+    const newQuantity = Number(productToSell.quantity) - Number(quantity)
+    if(newQuantity<0) return res.status(400).send('You cannot sell more than you have.')
+    const updateStatus = await editProductdb(productId,{quantity:newQuantity})
+    const transaction = {
+        userId,
+        productId:productToSell.productId,
+        addedQuantity:0,
+        subtractedQuantity:Number(quantity),
+        soldQuantity:Number(quantity),
+        createdAt: Date.now()
+      }
+    await addTransactiondb(transaction)
+    return res.status(200).send(`${quantity} ${productToSell.name} sold successfully!!!`)
+  } catch(err){
+    return res.status(500).send(err.message)
+  }
+}
+module.exports = {addProduct,getProducts,deleteProduct,updateProduct,sellProduct}
